changed duration from long to int

+++ src/foam/nanos/crunch/Capability.js
@@ -87,7 +87,7 @@ foam.CLASS({
     },
     {
       name: 'duration',
-      class: 'Long',
+      class: 'Int',
       documentation: `To be used in the case where expiry is duration-based.
       The UserCapabilityJunction object will have its expiry configured to a DateTime based on the lower value of the two, expiry and duration`
     },
